Fixed selenium tests
