+++ src/app/build.gradle
@@ -97,7 +97,7 @@ task generateTestData() {
             commandLine 'docker', 'run', '--rm', '--entrypoint', 'create_orderly_demo.sh', '-u', uid, '-v', "$projectDir:/orderly", '-w', "/orderly", image, "."
         }
         
-        String owMigrateImage = "vimc/orderlyweb-migrate:master"
+        String owMigrateImage = "vimc/orderlyweb-migrate:mrc-1455"
         println("Pulling " + owMigrateImage)
         exec {
             commandLine 'docker', 'pull', owMigrateImage
