+++ src/customConfigTests/src/test/kotlin/org/vaccineimpact/orderlyweb/customConfigTests/GuestUserTests.kt
@@ -1,9 +1,12 @@
 package org.vaccineimpact.orderlyweb.customConfigTests
 
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.Before
 import org.junit.Test
 import org.openqa.selenium.By
 import org.openqa.selenium.support.ui.ExpectedConditions
+import org.vaccineimpact.orderlyweb.db.JooqContext
+import org.vaccineimpact.orderlyweb.db.Tables
 import org.vaccineimpact.orderlyweb.db.repositories.OrderlyAuthorizationRepository
 import org.vaccineimpact.orderlyweb.models.Scope
 import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
@@ -11,10 +14,26 @@ import org.vaccineimpact.orderlyweb.test_helpers.insertReport
 
 class GuestUserTests: SeleniumTest() {
 
+    private fun login()
+    {
+       wait.until(ExpectedConditions.presenceOfElementLocated(By.id("login_field")))
+       loginWithGithub()
+    }
+
+    @Before
+    fun setUp()
+    {
+        JooqContext().use {
+           it.dsl.update(Tables.ORDERLYWEB_SETTINGS)
+                   .set(Tables.ORDERLYWEB_SETTINGS.AUTH_ALLOW_GUEST, true)
+                   .execute()
+        }
+    }
+
     @Test
     fun `guest user can access homepage`() {
+        startApp("auth.provider=github")
 
-        startApp("auth.allow_guest=true")
         driver.get(RequestHelper.webBaseUrl)
 
         val header = driver.findElement(By.cssSelector("h1"))
@@ -28,7 +47,7 @@ class GuestUserTests: SeleniumTest() {
     @Test
     fun `guest user permisisons are updated immediately`() {
 
-        startApp("auth.allow_guest=true")
+        startApp("auth.provider=github")
         driver.get(RequestHelper.webBaseUrl)
 
         val header = driver.findElement(By.cssSelector("h1"))
@@ -61,9 +80,9 @@ class GuestUserTests: SeleniumTest() {
 
         insertReport("testreport", "20170103-143015-1234abcd")
         insertReport("testreport2", "20180103-143015-1234abcd")
-        addUserWithPermissions(listOf(ReifiedPermission("reports.read", Scope.Global())))
+        addUserWithPermissions(listOf(ReifiedPermission("reports.read", Scope.Global())), "notarealemail")
 
-        startApp("auth.allow_guest=true")
+        startApp("auth.provider=github")
         driver.get(RequestHelper.webBaseUrl)
 
         // guest user should not see any reports in report table
@@ -76,12 +95,7 @@ class GuestUserTests: SeleniumTest() {
         wait.until(ExpectedConditions.presenceOfElementLocated(By.className("login-link")))
         driver.findElement(By.className("login-link")).click()
 
-        // should be directed to login provider, i.e. Montagu in this case
-        wait.until(ExpectedConditions.presenceOfElementLocated(By.name("email")))
-
-        driver.findElement(By.name("email")).sendKeys("test.user@example.com")
-        driver.findElement(By.name("password")).sendKeys("password")
-        driver.findElement(By.id("login-button")).click()
+        login()
 
         // should be returned to the site as logged in user
         wait.until(ExpectedConditions.presenceOfElementLocated(By.className("site-title")))
@@ -89,6 +103,7 @@ class GuestUserTests: SeleniumTest() {
         // user should see logout link and should no longer see login link
         assertThat(driver.findElements(By.className("login")).count()).isEqualTo(0)
         assertThat(driver.findElements(By.className("logout")).count()).isEqualTo(1)
+        assertThat(driver.findElement(By.className("logout")).text).isEqualTo("Logged in as notarealemail | Logout")
 
         // user should see reports in report table
         rows = driver.findElements(By.cssSelector("table.dataTable tbody tr"))
